README: fix the slightly misleading doc
`make proto` command runs protocol generation based on the tools
installed on your local machine, which is not what really expected,
because tools can be outdated.

`make` calls a `make proto` from a container, so all tools are of
the correct versions.

This patch changes `make proto` to just `make`.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>