Allow to use switch NI ports as VLAN access ports
The main use-case of switch NI with multiple ports is not actually to
bridge multiple L2 switches for redundancy and run STP to avoid loops,
but to allow connecting external end devices (e.g. sensors) directly
into the same L2 segment as the applications running on the edge node.
With VLAN-enabled Switch NI, the user should be able to use a physical
network port attached to Switch NI as a VLAN access port (by default
all switch NI ports are configured as trunks).

Signed-off-by: Milan Lenco <milan@zededa.com>